Example repo that demonstrates dex2oat failure when app contains several hundred modules.
When compiling a debug variant of the app, AGP does not perform dex
merging on project modules. This means that there will be a single dex
per module included in the app. dex2oat has a limit of processing 512
dex files in an app and will throw an exception causing the apk to run
in an interpretted mode rather than the oatified native code for the
device.